task: Move hooks folder
