--- apps/client/app/hooks/fetcher.ts
@@ -1,2 +1,0 @@
-export const fetcher = (route: string) =>
-  fetch(`http://localhost:3100/api/${route}`).then((res) => res.json());
