+++ apps/client/app/parent-items/page.tsx
@@ -1,48 +1,16 @@
 'use client';
-import type { Item, ParentItem } from '../../types';
-import { useParentItems } from '../hooks/useParentItems';
+import { useParentItems } from '../../hooks/useParentItems';
 import Spinner from '../../components/common/Spinner';
+import ParentItemsTable from '../../components/pages/ParentItemsTable';
 
 export default function Index() {
-  const { parentItems, isLoading, error } = useParentItems();
+  // TODO: Add error handling
+  const { parentItems, isLoading } = useParentItems();
 
   return (
     <>
       <h1 className="text-2xl">Parent Items</h1>
-      {isLoading ? (
-        <Spinner />
-      ) : (
-        <table className="border-collapse table-auto w-full text-sm">
-          <thead>
-            <tr>
-              <th className="border-b border-slate-100 dark:border-slate-700 p-4 pl-8 text-slate-500 dark:text-slate-400 text-left">
-                Parent Item
-              </th>
-              <th className="border-b border-slate-100 dark:border-slate-700 p-4 pl-8 text-slate-500 dark:text-slate-400 text-left">
-                Items
-              </th>
-            </tr>
-          </thead>
-          <tbody className="bg-white dark:bg-slate-800">
-            {parentItems?.map((parentItem: ParentItem) => (
-              <tr key={parentItem.id}>
-                <td className="border-b border-slate-100 dark:border-slate-700 p-4 pl-8 text-slate-500 dark:text-slate-400">
-                  {parentItem.name}
-                </td>
-                <td className="border-b border-slate-100 dark:border-slate-700 p-4 pl-8 text-slate-500 dark:text-slate-400">
-                  <ul>
-                    {parentItem.items.map((item: Item) => (
-                      <li key={item.id}>
-                        {item.name} ({item.sku})
-                      </li>
-                    ))}
-                  </ul>
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      )}
+      {isLoading ? <Spinner /> : <ParentItemsTable parentItems={parentItems} />}
     </>
   );
 }
